Move permission prompt to settings page
