+++ background/handle-permissions.js
@@ -1,24 +1,24 @@
 import changeAddonState from "./imports/change-addon-state.js";
 import { getMissingOptionalPermissions } from "./imports/util.js";
 
-const onPermissionsRevoked = ({ isStartup }) => {
+const onPermissionsRevoked = ({ ask }) => {
   console.error("Site access is not granted.");
-  if (!isStartup) {
+  if (ask) {
     chrome.tabs.create({
       active: true,
-      url: "/webpages/settings/permissions.html",
+      url: "/webpages/settings/index.html",
     });
   }
 };
 
-const checkSitePermissions = (sendResponse, { isStartup }) => {
+const checkSitePermissions = (sendResponse, { ask }) => {
   chrome.permissions.contains(
     {
       origins: chrome.runtime.getManifest().host_permissions.filter((url) => url.startsWith("https://")),
     },
     (hasPermissions) => {
       if (!hasPermissions) {
-        onPermissionsRevoked({ isStartup });
+        onPermissionsRevoked({ ask });
       }
       sendResponse(hasPermissions);
     }
@@ -38,16 +38,19 @@ const checkOptionalPermissions = () => {
 
 chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
   if (request !== "checkPermissions") return;
-  checkSitePermissions(sendResponse, { isStartup: false });
+  // When the popup checks for permissions, redirect to the settings page if they are missing.
+  // When the settings page checks, just send a response; it does the asking.
+  const isFromPopup = sender.url === chrome.runtime.getURL(chrome.runtime.getManifest().action.default_popup);
+  checkSitePermissions(sendResponse, { ask: isFromPopup });
   return true;
 });
 
 chrome.permissions.onRemoved?.addListener(() => {
-  checkSitePermissions(() => {}, { isStartup: false });
+  checkSitePermissions(() => {}, { ask: true });
   checkOptionalPermissions();
 });
 
-checkSitePermissions(() => {}, { isStartup: true });
+checkSitePermissions(() => {}, { ask: false });
 checkOptionalPermissions();
 
 chrome.runtime.onInstalled.addListener((details) => {
@@ -55,6 +58,6 @@ chrome.runtime.onInstalled.addListener((details) => {
     // If the user just installed the extension, we do not consider this startup,
     // it's fine to open a new tab if needed in this case.
     // This happens on Firefox when loading the extension as temporary.
-    checkSitePermissions(() => {}, { isStartup: false });
+    checkSitePermissions(() => {}, { ask: true });
   }
 });
