--- webpages/settings/permissions.html
@@ -1,46 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title></title>
-    <script src="../set-lang.js" type="module"></script>
-    <link rel="icon" href="../../images/icon.png" id="favicon" />
-    <link rel="stylesheet" href="style.css" />
-    <link href="../../libraries/thirdparty/Sora.css" rel="stylesheet" />
-    <link href="../../libraries/thirdparty/Roboto.css" rel="stylesheet" />
-    <script src="../../libraries/thirdparty/vue.js" defer></script>
-    <script src="permissions.js" type="module"></script>
-    <style>
-      /*
-       * Imports are here to attempt to speed up load. Also, it means licenses.html loads fewer unused styles, and this file doesn't get licenses.html's styles! In order of when it will be noticed/needed on the page
-      */
-      @import url("../styles/colors.css");
-      @import url("../styles/components/buttons.css");
-    </style>
-  </head>
-  <body class="permissions">
-    <div class="navbar">
-      <a href="./index.html"
-        ><img src="../../images/icon-transparent.svg" class="logo header-button" alt="Logo" draggable="false"
-      /></a>
-      <h1 v-cloak>{{ msg("permissions") }}</h1>
-    </div>
-    <div v-cloak class="main">
-      <div class="permissions-block">
-        <h2>{{ msg("enablePermissionsTitle") }}</h2>
-        <p>{{ msg("enablePermissionsDescription") }}</p>
-      </div>
-      <div class="permissions-steps">
-        <div>
-          <button id="permissionsBtn" class="large-button">{{ msg("enableButton") }}</button>
-          <p>{{ msg("enableButtonDescription") }}</p>
-        </div>
-        <div>
-          <img alt="{{ msg('permissionsScreenAlt') }}" id="screenshot" v-bind:src="screenshotPath" />
-          <p>{{ msg("clickAllowDescription") }}</p>
-        </div>
-      </div>
-    </div>
-  </body>
-</html>
